+++ R/read-utf.R
@@ -69,9 +69,9 @@ read_str <- function(con) {
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 read_str_raw <- function(con, n) {
   vec <- read_raw(con, n)
-  res <- rawToChar(vec)
+  do_eof_check(con, n, length(vec))
   
-  do_eof_check(con, n, nchar(res))
+  res <- rawToChar(vec)
   res
 }